+++ frontend/src/app/add-request/page.tsx
@@ -1,28 +1,5 @@
-'use client'
-
-import { MainLayout } from "@/components/layouts/MainLayout"
-import AddRequestContent from '@/components/add-request/AddRequestContent'
-import { useAuthStore } from '@/store/authStore'
-import { useModal } from '@/contexts/ModalContext'
-import { useEffect, useState } from 'react'
+import { AddRequestPageClient } from './AddRequestPageClient';
 
 export default function AddRequestPage() {
-  const { isAuthenticated } = useAuthStore()
-  const { openModal } = useModal()
-  const [hasShownLoginModal, setHasShownLoginModal] = useState(false)
-
-  useEffect(() => {
-    if (!isAuthenticated && !hasShownLoginModal) {
-      openModal('signIn')
-      setHasShownLoginModal(true)
-    }
-  }, [isAuthenticated, openModal, hasShownLoginModal])
-
-  if (!isAuthenticated) return null
-
-  return (
-    <MainLayout>
-      <AddRequestContent />
-    </MainLayout>
-  )
+  return <AddRequestPageClient />;
 }
