+++ frontend/src/app/admin/locations/edit/EditLocationPageClient.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import EditLocation from '@/components/admin/dashboard/EditLocation'
+
+export function EditLocationPageClient() {
+  return <EditLocation />
+}
