+++ frontend/src/app/admin/verification-requests/page.tsx
@@ -1,15 +1,5 @@
-"use client"
-
-import { useEffect } from "react"
-import { useRouter } from "next/navigation"
+import { VerificationRequestsPageClient } from './VerificationRequestsPageClient';
 
 export default function VerificationRequestsPage() {
-  const router = useRouter()
-
-  // Redirect to the users page with the verification tab active
-  useEffect(() => {
-    router.replace("/admin/users?tab=verification")
-  }, [router])
-
-  return null
+  return <VerificationRequestsPageClient />;
 }
