+++ frontend/src/app/community/CommunityPageClient.tsx
@@ -1,0 +1,35 @@
+'use client'
+
+import { MainLayout } from "@/components/layouts/MainLayout";
+import CommunityFeed from "@/components/community/communityFeed";
+import CommunityWritePost from "@/components/community/communityWritePost";
+import { useState } from 'react';
+import { CommunityPost } from '@/types';
+
+export function CommunityPageClient() {
+  const [isWritingPost, setIsWritingPost] = useState(false);
+  const [newPost, setNewPost] = useState<CommunityPost | null>(null);
+
+  const handlePostSubmit = (createdPost: CommunityPost) => {
+    setNewPost(createdPost);
+    setIsWritingPost(false);
+  };
+
+  return (
+    <MainLayout>
+      <div className="container py-8">
+        {isWritingPost ? (
+          <CommunityWritePost 
+            onCancel={() => setIsWritingPost(false)}
+            onSubmitSuccess={handlePostSubmit}
+          />
+        ) : (
+          <CommunityFeed 
+            onWritePost={() => setIsWritingPost(true)} 
+            newPost={newPost}
+          />
+        )}
+      </div>
+    </MainLayout>
+  );
+}
