+++ frontend/src/app/community/page.tsx
@@ -1,37 +1,10 @@
-"use client"
-
-import { MainLayout } from "@/components/layouts/MainLayout";
-import CommunityFeed from "@/components/community/communityFeed";
-import CommunityWritePost from "@/components/community/communityWritePost";
-import { useState } from 'react';
-import { CommunityPost } from '@/types';
+import { Suspense } from 'react';
+import { CommunityPageClient } from './CommunityPageClient';
 
 export default function CommunityPage() {
-  const [isWritingPost, setIsWritingPost] = useState(false);
-  const [newPost, setNewPost] = useState<CommunityPost | null>(null);
-
-  const handlePostSubmit = (createdPost: CommunityPost) => {
-    setNewPost(createdPost);
-    setIsWritingPost(false);
-  };
-
   return (
-    <MainLayout>
-      <div className="container py-8">
-        {isWritingPost ? (
-          <CommunityWritePost 
-            onCancel={() => setIsWritingPost(false)}
-            onSubmitSuccess={handlePostSubmit}
-          />
-        ) : (
-          <CommunityFeed 
-            onWritePost={() => setIsWritingPost(true)} 
-            newPost={newPost}
-          />
-        )}
-      </div>
-    </MainLayout>
+    <Suspense fallback={<div>Loading...</div>}>
+      <CommunityPageClient />
+    </Suspense>
   );
 }
-
-// Using CommunityPost interface imported from @/types
