+++ frontend/src/app/contactUs/page.tsx
@@ -1,13 +1,16 @@
 // C:\Users\MC\Desktop\Donation\donation-platform\frontend\src\app\contactUs\page.tsx
+import { Suspense } from "react";
 import { MainLayout } from "@/components/layouts/MainLayout";
 import ContactUs from "@/components/contactUs/contactUs";
 import { ContactInfo } from "@/components/contactUs/contactInfo";
 
 export default function ContactPage() {
   return (
-    <MainLayout>
-      <ContactInfo />
-      <ContactUs />
-    </MainLayout>
+    <Suspense fallback={<div>Loading...</div>}>
+      <MainLayout>
+        <ContactInfo />
+        <ContactUs />
+      </MainLayout>
+    </Suspense>
   );
 }
