--- frontend/src/app/test/image-gallery/page.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import ImageGalleryTest from '@/components/test/ImageGalleryTest'
-
-export default function ImageGalleryTestPage() {
-  return <ImageGalleryTest />
-}
